% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{add_agg_medians}
\alias{add_agg_medians}
\title{Add aggregate medians}
\usage{
add_agg_medians(df, fill_gaps, data_dir)
}
\arguments{
\item{df}{data frame from \code{fg_pip()} or \code{rg_pip()}}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{data_dir}{character: Directory path of auxiliary data (lkup$data_root)}
}
\value{
data.table
}
\description{
For lineup years all medians are set to NA; for survey years the existing
median is preferred and missing values are filled from the spr table.
}
\keyword{internal}
