% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{add_dist_stats}
\alias{add_dist_stats}
\title{Add pre-computed distributional stats (new pathway)}
\usage{
add_dist_stats(df, lkup, fill_gaps)
}
\arguments{
\item{df}{data.table: Data frame of poverty statistics}

\item{lkup}{list: lookup object containing dist_stats and lineup_dist_stats}

\item{fill_gaps}{logical: whether lineup-year estimates are being used}
}
\value{
data.table
}
\description{
Add pre-computed distributional stats (new pathway)
}
