% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_dist_stats_old}
\alias{add_dist_stats_old}
\title{Add pre-computed distributional stats}
\usage{
add_dist_stats_old(df, dist_stats)
}
\arguments{
\item{df}{data.table: Data frame of poverty statistics}

\item{dist_stats}{data.table: Distributional stats lookup}
}
\value{
data.table
}
\description{
Add pre-computed distributional stats
}
