% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{add_distribution_type}
\alias{add_distribution_type}
\title{Classify surveys by distribution type}
\usage{
add_distribution_type(df, lkup, fill_gaps)
}
\arguments{
\item{df}{data frame from \code{fg_pip()} or \code{rg_pip()}}

\item{lkup}{list: lookup table}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}
}
\value{
data.table
}
\description{
Uses framework data to classify each survey year as micro, group, imputed,
or mixed. For lineup years, mixed classification is applied when a reporting
year spans surveys of different types.
}
\keyword{internal}
