% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{add_spl}
\alias{add_spl}
\title{Add Shared Prosperity Line indicators}
\usage{
add_spl(df, fill_gaps, data_dir)
}
\arguments{
\item{df}{data frame inside \code{fg_pip()} or \code{rg_pip()}}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{data_dir}{character: Directory path of auxiliary data (lkup$data_root)}
}
\value{
data.table
}
\description{
Add Shared Prosperity Line indicators
}
\keyword{internal}
