% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{add_vars_out_of_pipeline}
\alias{add_vars_out_of_pipeline}
\title{Add all variables estimated outside the core pipeline}
\usage{
add_vars_out_of_pipeline(out, fill_gaps, lkup)
}
\arguments{
\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{lkup}{list: lookup table}
}
\value{
data.table (modified in-place)
}
\description{
Orchestrates the addition of SPL/SPR, Prosperity Gap, and distribution
type. Any future out-of-pipeline variables should be added here.
}
\keyword{internal}
