% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_agg_stats.R
\name{ag_average_poverty_stats}
\alias{ag_average_poverty_stats}
\title{Compute poverty statistics for aggregated data distribution.}
\usage{
ag_average_poverty_stats(df)
}
\arguments{
\item{area}{character: Area (Urban or Rural)}

\item{area_pop}{numeric: Total population per area.}
}
\value{
list
}
\description{
Compute poverty statistics for aggregated data distribution.
}
\keyword{internal}
