% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_lkup.R
\name{assert_lkup_field}
\alias{assert_lkup_field}
\title{Assert that a single field is present in a lkup list}
\usage{
assert_lkup_field(lkup, field)
}
\arguments{
\item{lkup}{list: The lookup object to check.}

\item{field}{character scalar: The field name to look for.}
}
\value{
Invisibly returns \code{lkup} if the field exists; aborts otherwise.
}
\description{
Assert that a single field is present in a lkup list
}
\examples{
\dontrun{
assert_lkup_field(lkup, "svy_lkup")
}
}
\keyword{internal}
