% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{assign_required_params}
\alias{assign_required_params}
\title{Assign PIP API required parameters if missing}
\usage{
assign_required_params(req)
}
\arguments{
\item{req}{list: plumber \code{req} object}
}
\value{
list
}
\description{
Assign PIP API required parameters if missing
}
