% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{assign_serializer}
\alias{assign_serializer}
\title{Helper function to return correct serializer}
\usage{
assign_serializer(format)
}
\arguments{
\item{format}{characer: Response format. Options are "json", "csv", or "rds"}
}
\value{
serializer function
}
\description{
Helper function to return correct serializer
}
