% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_cumsum.R
\name{build_pair_dict}
\alias{build_pair_dict}
\title{Dictionary for fast joins}
\usage{
build_pair_dict(lkup, fill_gaps = TRUE)
}
\arguments{
\item{lkup}{Lookup object containing refy_lkup and svy_lkup.}

\item{fill_gaps}{Logical, TRUE for lineup years, FALSE for survey years.}
}
\value{
data.table with dictionary for merges.

data.table with columns id, reporting_level, and code.
}
\description{
Dictionary for fast joins
}
\keyword{Build}
\keyword{Creates}
\keyword{a}
\keyword{and}
\keyword{by}
\keyword{data.table}
\keyword{dictionary}
\keyword{fast}
\keyword{for}
\keyword{id}
\keyword{internal}
\keyword{joins}
\keyword{merging}
\keyword{reporting_level.}
