% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-censor.R
\name{censor_stats}
\alias{censor_stats}
\title{Apply censor table to a data.table}
\usage{
censor_stats(df, censored_table)
}
\arguments{
\item{df}{data.table: Table to censor (must have a \code{tmp_id} column).}

\item{censored_table}{data.table: Censor table with columns \code{id} and
\code{statistic}.}
}
\value{
data.table
}
\description{
Removes rows flagged with \code{statistic == "all"} and sets individual
statistic columns to \code{NA} for partial censors.
}
\keyword{internal}
