% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp_logic.R
\name{check_inputs_pip_grp_logic}
\alias{check_inputs_pip_grp_logic}
\title{Check parameters of pip_grp_logic}
\usage{
check_inputs_pip_grp_logic(
  country,
  year,
  povline,
  popshare,
  group_by,
  welfare_type,
  reporting_level,
  lkup,
  debug,
  censor
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{group_by}{character: Will return aggregated values for predefined
sub-groups}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{debug}{logical: If TRUE poverty calculations from \code{wbpip} will run in
debug mode}

\item{censor}{logical: Triggers censoring of country/year statistics}
}
\value{
\code{invsible(TRUE)}
}
\description{
Check parameters of pip_grp_logic
}
