% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{check_param_chr}
\alias{check_param_chr}
\title{Check validity of a single query parameter}
\usage{
check_param_chr(values, valid_values)
}
\arguments{
\item{values}{character: query values}

\item{valid_values}{character: valid values}
}
\value{
logical
}
\description{
Check validity of a single query parameter
}
