% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{check_param_lgl}
\alias{check_param_lgl}
\title{Check validity of a single query parameter}
\usage{
check_param_lgl(value)
}
\arguments{
\item{values}{logical: query values}
}
\value{
logical
}
\description{
Check validity of a single query parameter
}
