% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{check_param_num}
\alias{check_param_num}
\title{Check validity of a single query parameter}
\usage{
check_param_num(value, valid_values)
}
\arguments{
\item{valid_values}{list: List of valid bounded values}

\item{values}{numeric: query values}
}
\value{
logical
}
\description{
Check validity of a single query parameter
}
