% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{check_parameters}
\alias{check_parameters}
\title{Check validity of query parameters}
\usage{
check_parameters(req, query_controls)
}
\arguments{
\item{req}{req: plumber request environment}

\item{query_controls}{list: List of valid values}
}
\value{
logical
}
\description{
Check validity of query parameters
}
