% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{citation_from_version}
\alias{citation_from_version}
\title{Return citation from the version}
\usage{
citation_from_version(version)
}
\arguments{
\item{version}{character vector of data version}
}
\value{
character. Text containing citation for the version passed.
}
\description{
Return citation from the version
}
