% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_data.R
\name{clean_api_data}
\alias{clean_api_data}
\title{Temporary function that takes care of basic data cleaning and validation}
\usage{
clean_api_data(data_folder_root)
}
\arguments{
\item{data_folder_root}{character: Path to data folder relative to the location
of the plumber.R file}
}
\value{
list
}
\description{
Temporary function that takes care of basic data cleaning and validation
}
