% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear a cachem disk cache}
\usage{
clear_cache(cd)
}
\arguments{
\item{cd}{A \code{cachem::cache_disk()} object}
}
\value{
Named list with \code{$status} (\code{"success"} or
\code{"error"}) and \code{$msg}.
}
\description{
Resets the cache and reports success or failure.
}
\keyword{internal}
