% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_logs}
\alias{clear_logs}
\title{Clear logs
Clear logs is available}
\usage{
clear_logs(logs_path)
}
\arguments{
\item{logs_path}{character: Path to log file}
}
\value{
list
}
\description{
Clear logs
Clear logs is available
}
