% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fgt_new.R
\name{compute_fgt}
\alias{compute_fgt}
\title{Efficient FGT calculation for vectors (No data.table)}
\usage{
compute_fgt(w, wt, povlines)
}
\arguments{
\item{w}{character: welfare variable name}

\item{wt}{character: weight variable name}

\item{povlines}{double: vector with poverty lines}
}
\value{
data.table with estimates poverty estimates
}
\description{
Efficient FGT calculation for vectors (No data.table)
}
\keyword{internal}
