% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_pip_old.R
\name{compute_fgt_dt_old}
\alias{compute_fgt_dt_old}
\title{Title}
\usage{
compute_fgt_dt_old(dt, welfare, weight, povlines)
}
\arguments{
\item{dt}{data frame with \code{welfare} and \code{weight} columns}

\item{welfare}{character: welfare variable name}

\item{weight}{character: weight variable name}

\item{povlines}{double: vector with poveryt lines}
}
\value{
data.table with estimates poverty estimates
}
\description{
Title
}
\keyword{internal}
