% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{create_etag_header}
\alias{create_etag_header}
\title{create_etag_header}
\usage{
create_etag_header(req)
}
\arguments{
\item{req}{R6 object: Plumber API request}
}
\value{
character
}
\description{
helper function that creates a unique hash of code + data
this hash value will be used as the value of the etag header
to facilitate caching of PIP API responses
}
