% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{create_full_list}
\alias{create_full_list}
\title{Create full list for fg data load, not including country-years in cache}
\usage{
create_full_list(metadata)
}
\arguments{
\item{metadata}{data table from subset_lkup()$lkup}
}
\value{
data.table
}
\description{
Create full list for fg data load, not including country-years in cache
}
\keyword{internal}
