% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{create_return_cols}
\alias{create_return_cols}
\title{helper function to create a list of return columns for various pipapi functions}
\usage{
create_return_cols(...)
}
\arguments{
\item{...}{Named vectors of columns to be returned}
}
\value{
list
}
\description{
helper function to create a list of return columns for various pipapi functions
}
