% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{create_vintage_pattern_call}
\alias{create_vintage_pattern_call}
\title{create vintage call to be parsed into \code{get_vintage_pattern_regex()}}
\usage{
create_vintage_pattern_call(vintage_pattern = NULL)
}
\arguments{
\item{vintage_pattern}{either NULL, chracter with regex or list of arguments
for \code{get_vintage_pattern_regex()}}
}
\value{
list to be parses t \code{get_vintage_pattern_regex()}
}
\description{
create vintage call to be parsed into \code{get_vintage_pattern_regex()}
}
\examples{
vintage_pattern <- NULL
create_vintage_pattern_call(vintage_pattern)

vintage_pattern <- list("r.*", "", "^hjkhj\\\\.d")
create_vintage_pattern_call(vintage_pattern)

vintage_pattern <- c("r.*", "", "^hjkhj\\\\.d")
create_vintage_pattern_call(vintage_pattern)

vintage_pattern <- c(vintage_pattern = "r.*", test_regex = "", int_regex =  "^hjkhj\\\\.d")
create_vintage_pattern_call(vintage_pattern)
}
