% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_cumsum.R
\name{decode_pairs}
\alias{decode_pairs}
\title{Decode integer code to (id, reporting_level) labels}
\usage{
decode_pairs(
  DT,
  dict,
  code_col = "id_rl",
  id_col = "id",
  level_col = "reporting_level",
  keep_code = FALSE,
  add_true_vars = TRUE,
  verbose = 0L
)
}
\arguments{
\item{DT}{data.table to decode.}

\item{dict}{data.table from build_pair_dict().}

\item{code_col}{Name of code column in DT.}

\item{id_col}{Name of id column in dict.}

\item{level_col}{Name of reporting level column in dict.}

\item{keep_code}{Logical, keep code column if TRUE.}

\item{add_true_vars}{Logical, add country_code and reporting_year columns and remove id.}

\item{verbose}{Integer, verbosity level.}
}
\value{
data.table with id and reporting_level columns added.
}
\description{
Joins a data.table with a dictionary to recover id and reporting_level columns from integer codes.
Used after FGT calculations to restore human-readable labels.
}
\keyword{internal}
