% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_cumsum.R
\name{.get_pairs_df}
\alias{.get_pairs_df}
\title{Extract id and level columns as a data.frame}
\usage{
.get_pairs_df(X, id_col, level_col)
}
\arguments{
\item{X}{data.table or data.frame}

\item{id_col}{Name of id column}

\item{level_col}{Name of level column}
}
\value{
data.frame with id and level columns
}
\description{
Extract id and level columns as a data.frame
}
\keyword{internal}
