% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fgt_new.R
\name{DT_fgt_by_rl}
\alias{DT_fgt_by_rl}
\title{compute FGT using indices by reporting level}
\usage{
DT_fgt_by_rl(x, y, nx)
}
\arguments{
\item{x}{data.table from lt list, with welfare and weight vectors}

\item{y}{list of indices for each reporting level}

\item{nx}{name of data table. Usuall country code and year in the form "CCC_YYYY"}
}
\value{
data.table with FGT estimates by reporting level
}
\description{
This function is intended to be used inside \link{map_fgt}
}
\keyword{internal}
