% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_function_objects.R, R/pipapi-package.R
\docType{data}
\name{empty_response}
\alias{empty_response}
\title{Empty response schema}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 0 rows and 44 columns.

Data frame with 0 rows and 44 columns
}
\usage{
empty_response

data(empty_response)
}
\description{
A template for empty responses
}
\keyword{datasets}
