% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi-package.R
\docType{data}
\name{empty_response_cp_poverty}
\alias{empty_response_cp_poverty}
\title{List of two datasets \code{pov_trend} and \code{pov_mrv}}
\format{
A list with 2 dataframes
}
\usage{
data(empty_response_cp_poverty)
}
\description{
List of two datasets \code{pov_trend} and \code{pov_mrv}
}
\keyword{datasets}
