% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi-package.R
\docType{data}
\name{empty_response_fg_add}
\alias{empty_response_fg_add}
\title{Dataframe for fill gaps empty response with additional variables}
\format{
Data frame with 0 rows and 56 columns
}
\usage{
data(empty_response_fg_add)
}
\description{
Dataframe for fill gaps empty response with additional variables
}
\keyword{datasets}
