% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_cumsum.R
\name{encode_pairs}
\alias{encode_pairs}
\title{Encode pairs with integer code}
\usage{
encode_pairs(
  DT,
  dict,
  id_col = "id",
  level_col = "reporting_level",
  code_col = "id_rl",
  drop_labels = FALSE,
  strict = TRUE,
  verbose = 0L
)
}
\arguments{
\item{DT}{data.table to encode.}

\item{dict}{data.table from build_pair_dict().}

\item{id_col}{Name of id column.}

\item{level_col}{Name of reporting level column.}

\item{code_col}{Name of code column to write.}

\item{drop_labels}{Logical, drop id and level columns if TRUE.}

\item{strict}{Logical, error if any pairs are missing from dict.}

\item{verbose}{Integer, verbosity level.}
}
\value{
data.table with code column added.
}
\description{
Adds an integer code column to a data.table by joining with a dictionary.
}
\keyword{internal}
