% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-censor.R
\name{estimate_type_var}
\alias{estimate_type_var}
\title{Label regional estimates by estimate type}
\usage{
estimate_type_var(df, lkup)
}
\arguments{
\item{df}{data.table: Table to process.}

\item{lkup}{list: lkup value}
}
\description{
Classifies regional estimates as \code{"actual"}, \code{"projection"}, or
\code{"nowcast"} and censors specific stats where applicable.
}
\keyword{internal}
