% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{extract_data_dirs}
\alias{extract_data_dirs}
\title{Extract list of data sub-directories from main data directory}
\usage{
extract_data_dirs(data_dir, vintage_pattern)
}
\arguments{
\item{data_dir}{character: Path to the main data directory}

\item{vintage_pattern}{character: regex that identifies the name pattern of
vintage folders}
}
\value{
character
}
\description{
Extract list of data sub-directories from main data directory
}
