% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{extract_endpoint}
\alias{extract_endpoint}
\title{helper function to extract endpoint from req$PATH_INFO object}
\usage{
extract_endpoint(path)
}
\arguments{
\item{path}{character: Information returned by req$PATH_INFO}
}
\value{
character
}
\description{
helper function to extract endpoint from req$PATH_INFO object
}
