% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{extract_identity}
\alias{extract_identity}
\title{Return identity from the version of the data}
\usage{
extract_identity(version)
}
\arguments{
\item{version}{character vector of data version}
}
\value{
character vector of identity
}
\description{
Return identity from the version of the data
}
