% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{extract_ppp_date}
\alias{extract_ppp_date}
\title{Return the ppp date from the version of the data}
\usage{
extract_ppp_date(version)
}
\arguments{
\item{version}{Data version}
}
\value{
Date of ppp
}
\description{
Return the ppp date from the version of the data
}
