% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{extract_release_date}
\alias{extract_release_date}
\title{Return the release date from the version of the data}
\usage{
extract_release_date(version)
}
\arguments{
\item{version}{Data version}
}
\value{
Date of release
}
\description{
Return the release date from the version of the data
}
