% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{extract_versions}
\alias{extract_versions}
\title{Extract data versions from vector of data directory paths}
\usage{
extract_versions(data_dirs, version_length)
}
\arguments{
\item{data_dirs}{character: List of data directory paths}

\item{version_length}{integer: Number of character of version ID}
}
\value{
character
}
\description{
Extract data versions from vector of data directory paths
}
