% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{fg_assign_nas_values_to_dup_cols}
\alias{fg_assign_nas_values_to_dup_cols}
\title{Coerce variable causing potential duplicates to NAs}
\usage{
fg_assign_nas_values_to_dup_cols(df, cols)
}
\arguments{
\item{df}{data.table: Table of results created in \code{fg_pip()}}

\item{cols}{character: Columns with potential duplicate values}
}
\value{
data.table
}
\description{
Coerce variable causing potential duplicates to NAs
}
\keyword{internal}
