% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip_old.R
\name{fg_assign_nas_values_to_dup_cols_old}
\alias{fg_assign_nas_values_to_dup_cols_old}
\title{OLD: Coerce variable causing potential duplicates to NAs}
\usage{
fg_assign_nas_values_to_dup_cols_old(df, cols)
}
\arguments{
\item{df}{data.table: Table of results created in \code{fg_pip()}}

\item{cols}{character: Columns with potential duplicate values}
}
\value{
data.table
}
\description{
OLD: Coerce variable causing potential duplicates to NAs
}
