% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{fg_get_mean_median}
\alias{fg_get_mean_median}
\title{merge into fgt table the mean and median from dist stats table in lkup}
\usage{
fg_get_mean_median(fgt, lkup)
}
\arguments{
\item{fgt}{data,table with fgt measures}

\item{lkup}{lkup}
}
\value{
data.table with with fgt, mean and median
}
\description{
merge into fgt table the mean and median from dist stats table in lkup
}
\keyword{internal}
