% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{fg_standardize_cache_id}
\alias{fg_standardize_cache_id}
\title{Standardize cache_id format to avoid duplication of rows}
\usage{
fg_standardize_cache_id(cache_id, interpolation_id, reporting_level)
}
\arguments{
\item{cache_id}{character}

\item{interpolation_id}{character}

\item{reporting_level}{character}
}
\value{
character
}
\description{
Standardize cache_id format to avoid duplication of rows
}
\keyword{internal}
