% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip_old.R
\name{fg_standardize_cache_id_old}
\alias{fg_standardize_cache_id_old}
\title{OLD: Standardize cache_id format to avoid duplication of rows}
\usage{
fg_standardize_cache_id_old(cache_id, interpolation_id, reporting_level)
}
\arguments{
\item{cache_id}{character}

\item{interpolation_id}{character}

\item{reporting_level}{character}
}
\value{
character
}
\description{
OLD: Standardize cache_id format to avoid duplication of rows
}
