% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fgt_new.R
\name{fgt_watts_cumsum}
\alias{fgt_watts_cumsum}
\title{compute fgt and watts using cumulative welfare rather than means}
\usage{
fgt_watts_cumsum(y, w, lines)
}
\arguments{
\item{y}{numeric welfare (sorted ascending within the subgroup)}

\item{w}{numeric weights (same order as y)}

\item{lines}{numeric vector of poverty lines}
}
\value{
Returns a data.table with columns: line, fgt0,fgt1,fgt2,watts
}
\description{
compute fgt and watts using cumulative welfare rather than means
}
