% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{fillin_list}
\alias{fillin_list}
\title{Populate a list from same-named objects in the parent frame}
\usage{
fillin_list(l, assign = TRUE)
}
\arguments{
\item{l}{list to populate. All names must exist in the parent frame.}

\item{assign}{logical: whether to assign the filled list back to the
parent frame variable of the same name.}
}
\value{
The populated list \code{l} (invisibly).
}
\description{
Fills named elements of \code{l} with the values of identically-named
objects found in the calling function's environment.
}
\examples{
l <- list(x = NULL,
y = NULL,
z = NULL)

x <-  2
y <-  "f"
z <- TRUE
fillin_list(l)
l
}
