% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{filter_md}
\alias{filter_md}
\title{Helper function to filter missing data table}
\usage{
filter_md(md, ctr_alt_agg, year)
}
\arguments{
\item{md}{data.frame: Table of countries with missing data}

\item{ctr_alt_agg}{character: Countries from alternate aggregates}

\item{year}{character: year}
}
\value{
data.table
}
\description{
Helper function to filter missing data table
}
