% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{format_error}
\alias{format_error}
\title{format_error
Format error if check_parameters() returns TRUE}
\usage{
format_error(param, valid_values)
}
\arguments{
\item{param}{character or numeric: Parsed parameters}

\item{valid_values}{character or numeric: Accepted values for this parameter}
}
\value{
list
}
\description{
format_error
Format error if check_parameters() returns TRUE
}
