% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_indicators.R
\name{get_additional_indicators}
\alias{get_additional_indicators}
\title{Add set of extra indicators to pip output}
\usage{
get_additional_indicators(dt)
}
\arguments{
\item{dt}{data.frame: country level out from PIP}
}
\value{
data.frame
}
\description{
Add set of extra indicators to pip output
}
\keyword{internal}
