% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_indicators.R
\name{get_additional_indicators_grp}
\alias{get_additional_indicators_grp}
\title{Add set of extra indicators to pip output in aggregate data.}
\usage{
get_additional_indicators_grp(dt)
}
\arguments{
\item{dt}{data.frame: global/regional level  from PIP-grp}
}
\value{
data.frame
}
\description{
Add set of extra indicators to pip output in aggregate data.
}
\keyword{internal}
