% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux_table.R
\name{get_aux_table}
\alias{get_aux_table}
\title{Return specified auxiliary data}
\usage{
get_aux_table(data_dir = NULL, table, long_format = FALSE)
}
\arguments{
\item{data_dir}{character: Data directory}

\item{table}{character: Name of auxiliary table}

\item{long_format}{logical: do you want data long format ? (default FALSE)}
}
\value{
data.frame
}
\description{
Return specified auxiliary data
}
\seealso{
\code{\link[=blocked_aux_tables]{blocked_aux_tables()}}

Other aux:
\code{\link[=blocked_aux_tables]{blocked_aux_tables()}},
\code{\link[=get_aux_table_ui]{get_aux_table_ui()}},
\code{\link[=reset_blocked_aux_tables]{reset_blocked_aux_tables()}}
}
\concept{aux}
