% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{get_caller_names}
\alias{get_caller_names}
\title{Get function names on the call stack}
\usage{
get_caller_names()
}
\value{
character vector of call names (invisibly)
}
\description{
Walks \code{sys.calls()} and returns a character vector of function name
strings. Handles \code{do.call()} specially by peeking at the next frame.
}
