% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{get_ctr_alt_agg}
\alias{get_ctr_alt_agg}
\title{Helper function to retrieve the required countries
needed to compute alternative aggregates requested by user
Get countries that belong to aggregates requested by the user that are NOT
official but alternative aggregates. We need to find out missing data
estimates only for those countries. For instance, if the user requested LAC
and AFE, we don't care about the the countries with missing data in the LAC
because their estimates are done implicitly. We DO care about the estimates
of the missing countries in AFE because we need the explicit SSA estimates.}
\usage{
get_ctr_alt_agg(user_alt_gt, user_alt_gt_code, user_alt_agg, cl)
}
\arguments{
\item{user_alt_gt}{character: Grouping type needed by user}

\item{user_alt_gt_code}{character: Grouping type code}

\item{user_alt_agg}{character: Alternate aggregates requested by user}

\item{cl}{data.frame: Countries lookup table}
}
\value{
character
}
\description{
Helper function to retrieve the required countries
needed to compute alternative aggregates requested by user
Get countries that belong to aggregates requested by the user that are NOT
official but alternative aggregates. We need to find out missing data
estimates only for those countries. For instance, if the user requested LAC
and AFE, we don't care about the the countries with missing data in the LAC
because their estimates are done implicitly. We DO care about the estimates
of the missing countries in AFE because we need the explicit SSA estimates.
}
