% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipdata.R
\name{get_dt_dist_stats}
\alias{get_dt_dist_stats}
\title{get data.table with distribution stats}
\usage{
get_dt_dist_stats(lt_att)
}
\arguments{
\item{lt_att}{list of attributes of lt list}
}
\value{
data.table
}
\description{
this is a loop over lt attributes
}
\keyword{internal}
