% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi-env.R
\name{get_from_pipapienv}
\alias{get_from_pipapienv}
\title{Get a value from .pipapienv}
\usage{
get_from_pipapienv(key)
}
\arguments{
\item{key}{A character string representing the key}
}
\value{
The value associated with the key in .pipapienv
}
\description{
Get a value from .pipapienv
}
\examples{
set_in_pipapienv("example_key", 42)
get_from_pipapienv("example_key") # returns 42
}
