% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{get_grp_to_compute}
\alias{get_grp_to_compute}
\title{Helper function to retrieve the country/year pairs to be computed}
\usage{
get_grp_to_compute(user_off_reg, md_off_reg, year, md_year)
}
\arguments{
\item{user_off_reg}{character: Official regions requested by user}

\item{md_off_reg}{character: Missing data for official regions}

\item{year}{character: Years}

\item{md_year}{character: Years with missing data}
}
\value{
data.frame
}
\description{
Helper function to retrieve the country/year pairs to be computed
}
