% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{get_impl_ctrs}
\alias{get_impl_ctrs}
\title{Helper function to retrieve the implicit country surveys present in both
alternative and official aggregates}
\usage{
get_impl_ctrs(user_gt, user_gt_code, user_aggs, ctrs)
}
\arguments{
\item{user_gt}{character: Grouping type}

\item{user_gt_code}{character: Grouping type code}

\item{user_aggs}{character: Aggregates selected by user}

\item{ctrs}{data.frame: Countries lookup table}
}
\value{
character
}
\description{
Helper function to retrieve the implicit country surveys present in both
alternative and official aggregates
}
