% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{get_logs}
\alias{get_logs}
\title{Helper function to retrieve most recent logs}
\usage{
get_logs(last_n = 20, log_file)
}
\arguments{
\item{last_n}{numeric: Last n lines of the log file}

\item{log_file}{character: Path to log file}
}
\value{
character
}
\description{
Helper function to retrieve most recent logs
}
