% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{get_md_vars}
\alias{get_md_vars}
\title{Helper function to retrieve variables needed to handle imputation of
missing data}
\usage{
get_md_vars(md, ctr_alt_agg, year, off_alt_agg, user_off_reg)
}
\arguments{
\item{md}{data.frame: Table of country/year with missing data}

\item{ctr_alt_agg}{character: Countries in alternative aggregate}

\item{year}{character: Years}

\item{off_alt_agg}{character: Instruction about how to handle official and
alternate aggregates}

\item{user_off_reg}{character: Official regions requested by user}
}
\value{
list
}
\description{
Helper function to retrieve variables needed to handle imputation of
missing data
}
