% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mean_median}
\alias{get_mean_median}
\title{merge into fgt table the mean and median from dist stats table in lkup}
\usage{
get_mean_median(fgt, lkup, fill_gaps)
}
\arguments{
\item{fgt}{data,table with fgt measures}

\item{lkup}{lkup}

\item{fill_gaps}{logical. whether to use lineup estimates}
}
\value{
data.table with with fgt, mean and median
}
\description{
merge into fgt table the mean and median from dist stats table in lkup
}
\keyword{internal}
