% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aux.R
\name{get_metaregion_table}
\alias{get_metaregion_table}
\title{Load metaregion table from aux data}
\usage{
get_metaregion_table(data_dir)
}
\arguments{
\item{data_dir}{character: Data directory}
}
\value{
data.table
}
\description{
Returns an empty data.table with the correct schema when no data is
available.
}
\keyword{internal}
