% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_values}
\alias{get_param_values}
\title{Get valid query parameter values
Get vector of accepted query parameter values}
\usage{
get_param_values(param)
}
\arguments{
\item{param}{character: Query parameter}
}
\description{
Get valid query parameter values
Get vector of accepted query parameter values
}
