% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aux.R
\name{get_pg_table}
\alias{get_pg_table}
\title{Load Prosperity Gap table from aux data}
\usage{
get_pg_table(data_dir, table = c("pg_svy", "pg_lnp"))
}
\arguments{
\item{data_dir}{character: Data directory}

\item{table}{character: one of \code{"pg_svy"} or \code{"pg_lnp"}}
}
\value{
data.table
}
\description{
Returns an empty data.table with the correct schema when no data is
available.
}
\keyword{internal}
