% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pip_version.R
\name{get_pip_version}
\alias{get_pip_version}
\title{Return the versions of the pip packages used for computations}
\usage{
get_pip_version(pip_package = c("pipapi"), lkup)
}
\arguments{
\item{lkup}{list: A list of lkup tables}

\item{pip_packages}{character: Custom packages powering the API}
}
\value{
list
}
\description{
Return the versions of the pip packages used for computations
}
