% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi-env.R
\name{get_pipapienv}
\alias{get_pipapienv}
\title{Get the entire .pipapienv environment}
\usage{
get_pipapienv()
}
\value{
The .pipapienv environment
}
\description{
Get the entire .pipapienv environment
}
\examples{
env <- get_pipapienv()
}
