% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_pip.R
\name{get_pov_estimates}
\alias{get_pov_estimates}
\title{get POV estiamates}
\usage{
get_pov_estimates(lt, povline, threshold = 500)
}
\arguments{
\item{lt}{list with data frames}

\item{povline}{double: vector with poverty lines}

\item{threshold}{integer: minimum length of X to use future_lapply (default 30)}
}
\value{
list
}
\description{
Internally it selects lapply or future_lapply
}
\keyword{internal}
