% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipdata.R
\name{get_rl_rows_single}
\alias{get_rl_rows_single}
\alias{get_rl_rows}
\title{extract rows indices}
\usage{
get_rl_rows_single(a)

get_rl_rows(lt_att)
}
\arguments{
\item{a}{list with attributes from lt}
}
\value{
names list with indices for reporting level
}
\description{
extract rows indices

apply get_rl_rows_single
}
\keyword{internal}
