% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aux.R
\name{get_spr_table}
\alias{get_spr_table}
\title{Load SPR table from aux data}
\usage{
get_spr_table(data_dir, table = c("spr_svy", "spr_lnp"))
}
\arguments{
\item{data_dir}{character: Data directory}

\item{table}{character: one of \code{"spr_svy"} or \code{"spr_lnp"}}
}
\value{
data.table
}
\description{
Returns an empty data.table with the correct schema when no data is
available.
}
\keyword{internal}
