% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aux.R
\name{get_svy_data}
\alias{get_svy_data}
\title{Read survey microdata}
\usage{
get_svy_data(svy_id, reporting_level, path)
}
\arguments{
\item{svy_id}{character: Survey ID (used only for naming output list elements)}

\item{reporting_level}{character: Geographical reporting level}

\item{path}{character: Path(s) to survey fst files}
}
\value{
Named list of data.tables, one per element of \code{path}
}
\description{
Reads welfare and weight columns from one or more fst files. For urban/rural
reporting levels the \code{area} column is used to subset rows.
}
\keyword{internal}
