% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_cumsum.R
\name{get_total_pop}
\alias{get_total_pop}
\title{Computes total population by group using the output of format_lfst().}
\usage{
get_total_pop(LDTg, dict)
}
\arguments{
\item{LDTg}{List from format_lfst() with DT and g objects.}

\item{dict}{data dictionary from build_pair_dict()}
}
\value{
data.table with total population by group.
}
\description{
Computes total population by group using the output of format_lfst().
}
\keyword{internal}
