% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{get_user_alt_gt}
\alias{get_user_alt_gt}
\title{Helper function to define user_alt_gt}
\usage{
get_user_alt_gt(user_gt, off_gt)
}
\arguments{
\item{user_gt}{character: Grouping type needed by user}

\item{off_gt}{character: Official grouping type}
}
\value{
character
}
\description{
Helper function to define user_alt_gt
}
