% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{get_user_x_code}
\alias{get_user_x_code}
\title{Helper function to define user_\{var\}_code}
\usage{
get_user_x_code(x)
}
\arguments{
\item{x}{character: Grouping type needed by user}
}
\value{
character
}
\description{
Helper function to define user_\{var\}_code
}
