% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{get_valid_aux_long_format_tables}
\alias{get_valid_aux_long_format_tables}
\title{Return auxiliary tables that support long_format=TRUE}
\usage{
get_valid_aux_long_format_tables()
}
\value{
character vector of valid table names
}
\description{
Return auxiliary tables that support long_format=TRUE
}
