% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{get_vintage_pattern_regex}
\alias{get_vintage_pattern_regex}
\title{Return regular expression needed for extracting data folders
Helper function to facilitate testing}
\usage{
get_vintage_pattern_regex()
}
\value{
list
}
\description{
Return regular expression needed for extracting data folders
Helper function to facilitate testing
}
