% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{id_valid_dirs}
\alias{id_valid_dirs}
\title{Identify valid data directories
Helper function to facilitate testing}
\usage{
id_valid_dirs(dirs_names, vintage_pattern)
}
\value{
logical
}
\description{
Identify valid data directories
Helper function to facilitate testing
}
