% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{ifel_isnull}
\alias{ifel_isnull}
\title{Efficient "if" "else" evaluation of null.}
\usage{
ifel_isnull(x, y)
}
\arguments{
\item{x}{object to evaluate}

\item{y}{in case x null. If X is not null, then x.}
}
\value{
object of class(x)
}
\description{
Efficient "if" "else" evaluation of null.
}
\keyword{internal}
