% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_empty}
\alias{is_empty}
\title{Test whether a vector is length zero and is not NULL}
\usage{
is_empty(x)
}
\arguments{
\item{x}{Vector to test}
}
\value{
logical. \code{TRUE} if \code{x} is empty but not \code{NULL}
}
\description{
Test whether a vector is length zero and is not NULL
}
\examples{
x <- vector()
is_empty(x)

y <- NULL
length(y)
is_empty(y)
}
