% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{is_forked}
\alias{is_forked}
\title{Helper function to determine whether an API call is compute intensive
and should be forked to a parallel process to avoid blocking the main
R process}
\usage{
is_forked(country, year, intensity_threshold = 40, include_year = TRUE)
}
\arguments{
\item{country}{character: selected countries}

\item{year}{character: selected years}

\item{intensity_threshold}{numeric: Number of selected country/year above which
the request will be considered intensive}

\item{include_year}{logical: Whether year selection should be included to determine
the intensity of the request}
}
\value{
logical
}
\description{
Helper function to determine whether an API call is compute intensive
and should be forked to a parallel process to avoid blocking the main
R process
}
