% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_pip.R
\name{is_future_plan_set}
\alias{is_future_plan_set}
\title{checks if a future plan is set (other than 'sequential')}
\usage{
is_future_plan_set()
}
\value{
logical value
}
\description{
checks if a future plan is set (other than 'sequential')
}
\keyword{internal}
