% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp_new.R
\name{list_code_column_values}
\alias{list_code_column_values}
\title{List values in each *_code column that match the country vector}
\usage{
list_code_column_values(dt, country)
}
\arguments{
\item{dt}{A data.table, typically lkup$aux_files$country_list.}

\item{country}{Character vector of country or region codes to match against *_code columns.}
}
\value{
A named list of unique values for each *_code column that match 'country'.
}
\description{
Returns a named list where each element is the vector of unique values in each *_code column
that are present in the provided 'country' vector.
}
\examples{
\dontrun{
dt <- lkup$aux_files$country_list
list_code_column_values(dt, c("USA", "EAP"))
}
}
