% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_releases.R
\name{list_data_releases}
\alias{list_data_releases}
\title{Return the list of PIP data releases}
\usage{
list_data_releases(root)
}
\arguments{
\item{root}{character: Path to root data folder}
}
\value{
character
}
\description{
Return the list of PIP data releases
}
