% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lkup.R
\name{lkup_filter}
\alias{lkup_filter}
\title{Filter a survey lookup table by country, year, welfare type, and reporting level}
\usage{
lkup_filter(
  lkup,
  country,
  year,
  valid_regions,
  reporting_level,
  welfare_type,
  data_dir
)
}
\arguments{
\item{country}{character: Country ISO3 codes or "ALL"}

\item{year}{integer or character: Reporting year(s), "ALL", or "MRV"}

\item{valid_regions}{character: Valid region codes from \code{lkup$query_controls}}

\item{reporting_level}{character: Requested reporting level}

\item{welfare_type}{character: Requested welfare type}

\item{data_dir}{character: Path to main data directory (\code{lkup$data_root})}
}
\description{
Filter a survey lookup table by country, year, welfare type, and reporting level
}
\keyword{internal}
