% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_pip_old.R
\name{load_data_list_old}
\alias{load_data_list_old}
\title{OLD: load survey year files and store them in a list}
\usage{
load_data_list_old(metadata)
}
\arguments{
\item{metadata}{data frame from \code{subset_lkup()}}
}
\value{
list with survey years data
}
\description{
OLD: load survey year files and store them in a list
}
\keyword{internal}
