% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipdata.R
\name{load_list_refy}
\alias{load_list_refy}
\title{load refy list}
\usage{
load_list_refy(input_list, path)
}
\arguments{
\item{input_list}{list. output from \link{create_full_list}}

\item{path}{character: directory path}
}
\value{
character vector
}
\description{
load refy list
}
\keyword{internal}
