% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{memo_norm}
\alias{memo_norm}
\title{momiose and normalize}
\usage{
memo_norm(f, cache)
}
\arguments{
\item{f}{function to normalize}

\item{cache}{object from \link[cachem:cache_disk]{cachem::cache_disk}}
}
\value{
memoised function
}
\description{
momiose and normalize
}
