% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_lapply}
\alias{my_lapply}
\title{selector of apply}
\usage{
my_lapply(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{optional arguments to \code{FUN}.}
}
\value{
list
}
\description{
selector of apply
}
\keyword{internal}
