% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{parse_parameters}
\alias{parse_parameters}
\title{parse_parameters}
\usage{
parse_parameters(params)
}
\arguments{
\item{params}{character: Query parameter to be parsed}
}
\value{
character
}
\description{
parse_parameters
}
