% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{parse_parameters_chr}
\alias{parse_parameters_chr}
\title{parse_parameters_chr}
\usage{
parse_parameters_chr(param)
}
\arguments{
\item{param}{character: Query parameter to be parsed}
}
\value{
character
}
\description{
parse_parameters_chr
}
