% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{parse_parameters_lgl}
\alias{parse_parameters_lgl}
\title{parse_parameters_lgl}
\usage{
parse_parameters_lgl(param)
}
\arguments{
\item{param}{character: Query parameter to be parsed}
}
\value{
integer
}
\description{
parse_parameters_lgl
}
