% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp.R
\name{pip_aggregate}
\alias{pip_aggregate}
\title{Calculate estiamtes for aggregates different to the official regional
aggregation}
\usage{
pip_aggregate(df, by = NULL)
}
\arguments{
\item{df}{data.table from \code{pip_fg()}}

\item{by}{character: Additional variable to use in \code{by} when doing the
aggregations. Default is \code{NULL}, but it should be use to include
aggregations variables}
}
\value{
data.table
}
\description{
Calculate estiamtes for aggregates different to the official regional
aggregation
}
