% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp.R
\name{pip_aggregate_by}
\alias{pip_aggregate_by}
\title{Aggregate by predefined groups}
\usage{
pip_aggregate_by(df, country = "ALL", return_cols = NULL)
}
\arguments{
\item{df}{data.frame: Response from \code{fg_pip_old()} or \code{rg_pip()}.}

\item{country}{character: Selected countries / regions}

\item{return_cols}{list: lkup$return_cols$pip_grp object. Controls returned
columns}
}
\description{
Aggregate by predefined groups
}
\keyword{internal}
