% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp_new.R
\name{pip_grp_new}
\alias{pip_grp_new}
\title{New way to estimate Aggregate data}
\usage{
pip_grp_new(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  lkup,
  censor = TRUE,
  additional_ind = FALSE
)
}
\description{
New way to estimate Aggregate data
}
