% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp_logic.R
\name{pip_grp_logic}
\alias{pip_grp_logic}
\title{Logic for computing new aggregate}
\usage{
pip_grp_logic(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  group_by = c("none", "wb"),
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  lkup,
  censor = TRUE,
  lkup_hash = lkup$cache_data_id$hash_pip_grp
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{group_by}{character: Will return aggregated values for predefined
sub-groups}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{censor}{logical: Triggers censoring of country/year statistics}

\item{lkup_hash}{character: hash of pip}
}
\value{
data.table
}
\description{
Logic for computing new aggregate
}
\examples{
\dontrun{
# Create lkups
}
}
