% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projection_var}
\alias{projection_var}
\title{projection variables}
\usage{
projection_var(df, censored_table)
}
\arguments{
\item{df}{data.table: Table to censor.}

\item{censored_table}{data.table: Censor table}
}
\description{
It also censors specific stats
}
\keyword{internal}
