% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_function_objects.R
\docType{data}
\name{reporting_level_list}
\alias{reporting_level_list}
\title{List of valid coverage values}
\format{
An object of class \code{character} of length 3.
}
\usage{
reporting_level_list
}
\description{
List of valid coverage values
}
\keyword{datasets}
