% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{ret_list}
\alias{ret_list}
\title{Store return values}
\usage{
ret_list(lret, x_name, x = get(x_name, envir = parent.frame()), assign = TRUE)
}
\arguments{
\item{lret}{list with return named objects predifined}

\item{x_name}{character: name of object to be added to \code{lret}. It must}

\item{x}{object to be added to lret}

\item{assign}{logical: whether to assign to parent frame. Default is TRUE}
}
\value{
invisible \code{lret} list
}
\description{
Store return values
}
\examples{
lf <- list(x = NULL,
y = 8,
w = "foo")

z <- "ocho"
x <- 2
w <- "nueve"
(ret_list(lf, "x"))
(ret_list(lf, "y", z))
(ret_list(lf, "w", assign = FALSE))
lf
}
