% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{return_correct_version}
\alias{return_correct_version}
\title{Return the version of the data}
\usage{
return_correct_version(
  version = NULL,
  release_version = NULL,
  ppp_version = NULL,
  identity = "PROD",
  versions_available
)
}
\arguments{
\item{version}{Data version. Defaults to most recent version. See api/v1/versions}

\item{release_version}{date when the data was published in YYYYMMDD format}

\item{ppp_version}{ppp year to be used}

\item{identity}{One of "PROD" (production), "INT" (internal) and "TEST"}

\item{versions_available}{character vector of all the versions available}
}
\value{
character
}
\description{
Return the version of the data
}
