% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_pip.R
\name{rg_pip}
\alias{rg_pip}
\title{Compute survey year stats}
\usage{
rg_pip(
  country,
  year,
  povline,
  popshare,
  welfare_type,
  reporting_level,
  ppp,
  lkup,
  con
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{ppp}{numeric: Custom Purchase Power Parity value}

\item{lkup}{list: A list of lkup tables}

\item{con}{duckdb connection object}
}
\value{
data.frame
}
\description{
Compute the main PIP poverty and inequality statistics for survey years.
}
\keyword{internal}
