% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lkup.R
\name{select_country}
\alias{select_country}
\title{Select country rows from a lookup table}
\usage{
select_country(lkup, keep, country, valid_regions)
}
\arguments{
\item{lkup}{list: A list of lkup tables}

\item{keep}{logical vector: current row selection mask}

\item{country}{character: Country ISO 3 codes}

\item{valid_regions}{character: List of valid region codes that can be used
for region selection}
}
\value{
logical vector
}
\description{
Helper function for \code{subset_lkup()}.
}
\keyword{internal}
