% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{select_off_alt_agg}
\alias{select_off_alt_agg}
\title{Helper function to identify how Official and Alternative regions should be
handled}
\usage{
select_off_alt_agg(user_gt, off_gt)
}
\arguments{
\item{user_gt}{character: Grouping type implicitly selected by user}

\item{off_gt}{character: Grouping type associated with Offical Regions}
}
\value{
character
}
\description{
Helper function to identify how Official and Alternative regions should be
handled
}
