% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{select_user_aggs}
\alias{select_user_aggs}
\title{Helper function to select correct Official Regions}
\usage{
select_user_aggs(country, off_reg, aggs)
}
\arguments{
\item{country}{character: User selected countries}

\item{off_reg}{character: Official region codes}

\item{aggs}{data.frame: Regions lookup table}
}
\value{
character
}
\description{
Helper function to select correct Official Regions
}
