% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lkup.R
\name{select_years}
\alias{select_years}
\title{Select year rows from a lookup table}
\usage{
select_years(lkup, keep, year, country, data_dir, valid_regions = NULL)
}
\arguments{
\item{lkup}{list: A list of lkup tables}

\item{keep}{logical vector: current row selection mask}

\item{year}{integer: Reporting year}

\item{country}{character: Country ISO 3 codes}

\item{data_dir}{character: directory path from lkup$data_root}

\item{valid_regions}{character: List of valid region codes that can be used
for region selection}
}
\value{
logical vector
}
\description{
Helper function for \code{subset_lkup()}.
}
\keyword{internal}
