% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi-env.R
\name{set_in_pipapienv}
\alias{set_in_pipapienv}
\title{Set a value in .pipapienv}
\usage{
set_in_pipapienv(key, value)
}
\arguments{
\item{key}{A character string representing the key}

\item{value}{The value to store in .pipapienv}
}
\value{
The assigned value (invisibly)
}
\description{
Set a value in .pipapienv
}
\examples{
set_in_pipapienv("example_key", 42)
}
