% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{sort_versions}
\alias{sort_versions}
\title{Sort available data folders
Helper function to facilitate unit testing}
\usage{
sort_versions(versions, prod_regex, int_regex, test_regex)
}
\arguments{
\item{versions}{character: vector of available versions}

\item{prod_regex}{character: Regex expression to identify production versions}

\item{int_regex}{character: Regex expression to identify internal versions}

\item{test_regex}{character: Regex expression to identify test versions}
}
\value{
character
}
\description{
Sort available data folders
Helper function to facilitate unit testing
}
