% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_api.R
\name{start_api}
\alias{start_api}
\title{Main function to launch the API}
\usage{
start_api(api_version = "v1", port = 80, host = "0.0.0.0")
}
\arguments{
\item{api_version}{character: API version to launch}

\item{port}{integer: Port}

\item{host}{character: Host}
}
\value{
plumber API
}
\description{
Main function to launch the API
}
