% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{subset_cp_rows}
\alias{subset_cp_rows}
\title{subset_cp_rows
TEMP function to select only national rows for
cases like CHN, IND etc.}
\usage{
subset_cp_rows(x)
}
\description{
subset_cp_rows
TEMP function to select only national rows for
cases like CHN, IND etc.
}
