% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_country_profile_optim.R
\name{ui_cp_charts_optim}
\alias{ui_cp_charts_optim}
\title{Country Profiles Charts}
\usage{
ui_cp_charts_optim(
  country = "AGO",
  povline = 1.9,
  pop_units = 1e+06,
  lkup,
  lkup_hash = lkup$cache_data_id$hash_ui_cp
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{povline}{numeric: Poverty line}

\item{pop_units}{numeric: Units used to express population numbers (default
to million)}

\item{lkup}{list: A list of lkup tables}

\item{lkup_hash}{character: hash of pip}
}
\value{
list
}
\description{
Provides numbers that will populate the country profile charts.
}
