% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_country_profile.R
\name{ui_cp_charts_single}
\alias{ui_cp_charts_single}
\title{Country profile charts for a single country}
\usage{
ui_cp_charts_single(country, povline, pop_units, lkup, pov_lkup = NULL)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{povline}{numeric: Poverty line}

\item{pop_units}{numeric: Units used to express population numbers (default
to million)}

\item{lkup}{list: A list of lkup tables}

\item{pov_lkup}{data.frame: Look up table for poverty charts.}
}
\value{
list
}
\description{
Country profile charts for a single country
}
\keyword{internal}
