% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_cp_download}
\alias{ui_cp_download}
\title{Country Profiles Key Indicators download}
\usage{
ui_cp_download(country = "AGO", year = "ALL", povline = 1.9, lkup)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}
}
\value{
list
}
\description{
Helper function to download Country Profile data
}
