% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_cp_download_single}
\alias{ui_cp_download_single}
\title{Country Profile Key Indicators download}
\usage{
ui_cp_download_single(country, povline = 1.9, lkup)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}
}
\value{
list
}
\description{
Helper function to download Country Profile data
}
\keyword{internal}
