% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_country_profile.R
\name{ui_cp_key_indicators}
\alias{ui_cp_key_indicators}
\title{Country Profiles Key Indicators}
\usage{
ui_cp_key_indicators(
  country = "AGO",
  povline = NULL,
  lkup,
  lkup_hash = lkup$cache_data_id$hash_ui_cp
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}

\item{lkup_hash}{character: hash of pip}
}
\value{
list
}
\description{
Provides numbers that will population for country profiles key indicators.
}
