% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_country_profile.R
\name{ui_cp_key_indicators_single}
\alias{ui_cp_key_indicators_single}
\title{Country Profile Key Indicators Single}
\usage{
ui_cp_key_indicators_single(country, povline, lkup)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}
}
\value{
list
}
\description{
Country profile key indicators for a single country.
}
\keyword{internal}
