% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_cp_ki_headcount}
\alias{ui_cp_ki_headcount}
\title{Provides numbers that will populate the country profiles key indicator for headcount}
\usage{
ui_cp_ki_headcount(country, povline, lkup)
}
\arguments{
\item{country}{character: Country code}

\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}
}
\value{
data.frame
}
\description{
Provides numbers that will populate the country profiles key indicator for headcount
}
\keyword{internal}
