% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_country_profile_optim.R
\name{ui_cp_poverty_charts_optim}
\alias{ui_cp_poverty_charts_optim}
\title{CP Poverty Charts}
\usage{
ui_cp_poverty_charts_optim(country, povline, pop_units, lkup)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{povline}{numeric: Poverty line}

\item{pop_units}{numeric: Units used to express population numbers (default
to million)}

\item{lkup}{list: A list of lkup tables}
}
\value{
list
}
\description{
Provides numbers that will populate the country profiles poverty charts
}
\keyword{internal}
