% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_home_page.R
\name{ui_hp_stacked}
\alias{ui_hp_stacked}
\title{Home Page Main Chart}
\usage{
ui_hp_stacked(povline = 1.9, lkup, lkup_hash = lkup$cache_data_id$hash_pip_grp)
}
\arguments{
\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}

\item{lkup_hash}{character: hash of pip}
}
\value{
data.table
}
\description{
Provides numbers that will populate the home page main chart.
}
