% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_svy_meta}
\alias{ui_svy_meta}
\title{Provides survey metadata that will populate the Data Sources page}
\usage{
ui_svy_meta(country = "all", lkup)
}
\arguments{
\item{country}{character: Country code}

\item{lkup}{list: A list of lkup tables}
}
\value{
data.frame
}
\description{
Provides survey metadata that will populate the Data Sources page
}
