% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{unnest_dt_longer}
\alias{unnest_dt_longer}
\title{Efficient unnest_longer for data.tables}
\usage{
unnest_dt_longer(tbl, cols)
}
\arguments{
\item{tbl}{a dataframe/tibble/data.table}

\item{cols}{one or more column names in \code{tbl} that contain list columns}
}
\value{
A longer data.table
}
\description{
An efficient alternative to \code{tidyr::unnest_longer()} that operates
directly on data.tables.
}
\examples{
\dontrun{
df <- data.frame(
 a = LETTERS[1:5],
 b = LETTERS[6:10],
 list_column1 = list(c(LETTERS[1:5]), "F", "G", "H", "I"),
 list_column2 = list(c(LETTERS[1:5]), "F", "G", "H", "K")
)
 unnest_dt_longer(df, grep("^list_column", names(df), value = TRUE))
}
}
