% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_cumsum.R
\name{update_pair_dict}
\alias{update_pair_dict}
\title{Update dictionary with new pairs}
\usage{
update_pair_dict(dict, DT, id_col = "id", level_col = "reporting_level")
}
\arguments{
\item{dict}{data.table dictionary from build_pair_dict().}

\item{DT}{data.table with id and reporting_level columns.}

\item{id_col}{Name of id column.}

\item{level_col}{Name of reporting level column.}
}
\value{
Updated data.table dictionary.
}
\description{
Appends new (id, reporting_level) pairs to the dictionary if needed.
}
\keyword{internal}
