% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{use_new_lineup_version}
\alias{use_new_lineup_version}
\title{Should the new lineup approach be used?}
\usage{
use_new_lineup_version(x)
}
\arguments{
\item{x}{A character vector where each element starts with an
8-digit date in the format \code{YYYYMMDD}.}
}
\value{
A logical vector: \code{TRUE} if the extracted date is
after May 1st, 2025, otherwise \code{FALSE}.
}
\description{
Check if the date in a string is more recent than May 2025
}
\details{
This function extracts the first 8 characters from an input string,
interprets them as a date in the format \code{YYYYMMDD}, and checks
whether this date is more recent than May 1st, 2025.
}
\examples{
use_new_lineup_version("20250401_2021_01_02_PROD") # FALSE
use_new_lineup_version("20250615_2021_01_02_PROD") # TRUE

}
