% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_years.R
\name{valid_years}
\alias{valid_years}
\title{Return available valid years}
\usage{
valid_years(data_dir)
}
\arguments{
\item{data_dir}{character: Path to the versioned data in main data directory}
}
\value{
numeric vector of valid years
}
\description{
Return available valid years
}
\examples{
valid_years('pip-fake-data/20200101_2011_01_01_PROD/')
}
