% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{validate_input_grouped_stats}
\alias{validate_input_grouped_stats}
\title{Validate grouped-stats endpoint input values}
\usage{
validate_input_grouped_stats(welfare, population, max_length = 100)
}
\arguments{
\item{welfare}{character: query values}

\item{population}{character: valid values}

\item{max_length}{integer: Max length of welfare vector}
}
\value{
list of two vectors welfare and population
}
\description{
Validate grouped-stats endpoint input values
}
