% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{version_dataframe}
\alias{version_dataframe}
\title{Return versions of the data available.}
\usage{
version_dataframe(versions)
}
\arguments{
\item{versions}{character: All available versions}
}
\value{
Dataframe with 4 columns, versions, release_version, ppp_version and identity
}
\description{
Return versions of the data available.
}
