% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_years.R
\name{wld_lineup_year}
\alias{wld_lineup_year}
\title{lineup year for the world}
\usage{
wld_lineup_year(data_dir)
}
\arguments{
\item{data_dir}{character: Path to the root directory}
}
\value{
numeric vector of length one of lineup year for the world
}
\description{
lineup year for the world
}
