% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_distribution_type}
\alias{add_distribution_type}
\title{Add Distribution type}
\usage{
add_distribution_type(df, lkup, fill_gaps)
}
\arguments{
\item{df}{data frame from \link{fg_pip} or \link{rg_pip}}

\item{lkup}{list: lookup table}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}
}
\value{
data.table
}
\description{
Add Distribution type
}
\keyword{internal}
