% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_spl}
\alias{add_spl}
\title{Add SPL indicators to either fg* or rg PIP output}
\usage{
add_spl(df, fill_gaps, data_dir)
}
\arguments{
\item{df}{data frame inside \link{fg_pip} or \link{rg_pip}}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{data_dir}{character: Directory path of auxiliary data. Usually
\code{lkup$data_root}}
}
\value{
data.table
}
\description{
Add SPL indicators to either fg* or rg PIP output
}
\keyword{internal}
