% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_vars_out_of_pipeline}
\alias{add_vars_out_of_pipeline}
\title{Add all the variables that are estimated outside the pipelines}
\usage{
add_vars_out_of_pipeline(out, fill_gaps, lkup)
}
\arguments{
\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{lkup}{list: lookup table}
}
\value{
data.table from pip or pip_grp functions.
}
\description{
This includes variables such as the SPL, SPR, PG, and distribution
type. Any other variables will be included here
}
\keyword{internal}
