% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_agg_stats.R
\name{ag_average_poverty_stats}
\alias{ag_average_poverty_stats}
\title{Compute poverty statistics for aggregated data distribution.}
\usage{
ag_average_poverty_stats(df)
}
\arguments{
\item{df}{data.frame: Survey data}
}
\value{
data.frame
}
\description{
Compute poverty statistics for aggregated data distribution.
}
