% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{available_versions}
\alias{available_versions}
\title{Sorted available PIP versions in data directory}
\usage{
available_versions(data_dir)
}
\arguments{
\item{data_dir}{character: data directory}
}
\value{
character vector of sorted available PIP versions in data directory
}
\description{
Sorted available PIP versions in data directory
}
