% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{censor_stats}
\alias{censor_stats}
\title{Censor stats}
\usage{
censor_stats(df, censored_table)
}
\arguments{
\item{df}{data.table: Table to censor.}

\item{censored_table}{data.table: Censor table}
}
\description{
Censor stats
}
\keyword{internal}
