% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{check_parameter}
\alias{check_parameter}
\title{Check validity of a single query parameter}
\usage{
check_parameter(values, valid_values, type)
}
\arguments{
\item{values}{vector: vector of values}

\item{valid_values}{vector: vector of valid values}

\item{type}{character: Type of value}
}
\value{
logical
}
\description{
Check validity of a single query parameter
}
