% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plumber.R
\name{check_parameters}
\alias{check_parameters}
\title{#' parse_parameters_int
#'
#' @param param character: Query parameter to be parsed
#'
#' @return integer
#' @export
#'
parse_parameters_int <- function(param) {}
\usage{
check_parameters(req, query_controls)
}
\arguments{
\item{req}{env: plumber request environment}

\item{query_controls}{list: List of valid values}
}
\value{
logical
}
\description{
if (!is.null(param)) {
param <- parse_parameters_chr(param)
param <- as.integer(param)
}
}
\details{

}
