% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cache
Clear cache directory if available}
\usage{
clear_cache(cd)
}
\arguments{
\item{cd}{A \code{cachem::cache_disk()} object}
}
\value{
list
}
\description{
Clear cache
Clear cache directory if available
}
\keyword{internal}
