% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fgt_new.R
\name{compute_fgt_dt}
\alias{compute_fgt_dt}
\title{Efficient FGT calculation for a data.table and vector of poverty lines}
\usage{
compute_fgt_dt(dt, welfare, weight, povlines, mean_and_med = FALSE)
}
\arguments{
\item{dt}{data frame with \code{welfare} and \code{weight} columns}

\item{welfare}{character: welfare variable name}

\item{weight}{character: weight variable name}

\item{povlines}{double: vector with poveryt lines}
}
\value{
data.table with estimates poverty estimates
}
\description{
Efficient FGT calculation for a data.table and vector of poverty lines
}
\keyword{internal}
