% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_vctr.R
\name{create_countries_vctr}
\alias{create_countries_vctr}
\title{Create countries vectors}
\usage{
create_countries_vctr(country, year, lkup)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{lkup}{list: A list of lkup tables}
}
\value{
a list of vectors with countries and regions code to be used in
\code{pip()} and \code{pip_grp()}
}
\description{
This functions selects the correct countries to be used in the aggregates
selected by the user, either official or alternative aggregates.
}
