% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{create_full_list}
\alias{create_full_list}
\title{Create full list for fg data load, not including country-years in cache}
\usage{
create_full_list(country, year, refy_lkup, data_present_in_master)
}
\arguments{
\item{country}{Country selected in \link{fg_pip} function}

\item{year}{Year/s selected in \link{fg_pip} function}

\item{refy_lkup}{reference year lkup table with full lineups and new method}

\item{data_present_in_master}{cache data}
}
\value{
list
}
\description{
Create full list for fg data load, not including country-years in cache
}
