% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{create_lkups}
\alias{create_lkups}
\title{Create look-up tables}
\usage{
create_lkups(data_dir, versions)
}
\arguments{
\item{data_dir}{character: Path to PIP data root folder.}

\item{versions}{character: Available data versions}
}
\value{
list
}
\description{
Create look-up tables that can be passed to \code{\link[=pip]{pip()}}.
}
