% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{create_versioned_lkups}
\alias{create_versioned_lkups}
\title{Create one list of lookups per data version}
\usage{
create_versioned_lkups(data_dir)
}
\arguments{
\item{data_dir}{character: Path to the main data directory}
}
\value{
list
}
\description{
Create one list of lookups per data version
}
