% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi.R
\docType{data}
\name{empty_response_fg}
\alias{empty_response_fg}
\title{Dataframe for fill gaps empty response}
\format{
Data frame with 0 rows and 53 columns
}
\usage{
data(empty_response_fg)
}
\description{
Dataframe for fill gaps empty response
}
\keyword{datasets}
