% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi.R
\docType{data}
\name{empty_response_grp}
\alias{empty_response_grp}
\title{Dataframe for grouped empty response}
\format{
Data frame with 0 rows and 12 columns
}
\usage{
data(empty_response_grp)
}
\description{
Dataframe for grouped empty response
}
\keyword{datasets}
