% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_type_ctr_lnp}
\alias{estimate_type_ctr_lnp}
\title{Add estimate_type var to lineup at the country level}
\usage{
estimate_type_ctr_lnp(out, lkup)
}
\arguments{
\item{out}{current data base}

\item{lkup}{lkup list}
}
\value{
out database with \code{estimate_type} variable
}
\description{
Add estimate_type var to lineup at the country level
}
\keyword{internal}
