% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_type_var}
\alias{estimate_type_var}
\title{projection variables}
\usage{
estimate_type_var(df, lkup)
}
\arguments{
\item{df}{data.table: Table to censor.}

\item{lkup}{lkup value}
}
\description{
It also censors specific stats
}
\keyword{internal}
