% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lkups.R
\name{extract_data_dirs}
\alias{extract_data_dirs}
\title{Extract list of data sub-directories from main data directory}
\usage{
extract_data_dirs(data_dir, version_length = 8)
}
\arguments{
\item{data_dir}{character: Path to main data directory}

\item{version_length}{integer: Number of character of version ID}
}
\value{
character
}
\description{
Extract list of data sub-directories from main data directory
}
