% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{fg_pip}
\alias{fg_pip}
\title{Compute imputed year stats}
\usage{
fg_pip(
  country,
  year,
  povline,
  popshare,
  welfare_type,
  reporting_level,
  ppp,
  ref_lkup,
  valid_regions,
  interpolation_list
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{ppp}{numeric: Custom Purchase Power Parity value}

\item{ref_lkup}{data.frame: Lookup table with necessary information to compute
inter/extra-polated values}

\item{valid_regions}{character: Vector of accepted country codes for regions}

\item{interpolation_list}{list: List of surveys to be used for computing
inter/extra-polated values}
}
\value{
data.frame
}
\description{
Compute the main PIP poverty and inequality statistics for imputed years.
}
\keyword{internal}
