% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{fg_pip}
\alias{fg_pip}
\title{Compute imputed year stats}
\usage{
fg_pip(
  country,
  year,
  povline,
  popshare,
  aggregate,
  welfare_type,
  reporting_level,
  ppp,
  lkup
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the poverty line}

\item{aggregate}{logical: If set to TRUE, will return aggregate results}

\item{welfare_type}{character: Welfare type.}

\item{reporting_level}{character: Geographical reporting level}

\item{ppp}{numeric: Custom Purchase Power Parity value}

\item{lkup}{list: A list of lkup tables}
}
\value{
data.frame
}
\description{
Compute the main PIP poverty and inequality statistics for imputed years.
}
\keyword{internal}
