% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pip.R
\name{fg_remove_duplicates}
\alias{fg_remove_duplicates}
\title{Remove duplicated rows created during the interpolation process}
\usage{
fg_remove_duplicates(
  df,
  cols = c("comparable_spell", "cpi", "display_cp", "gd_type", "interpolation_id",
    "path", "predicted_mean_ppp", "survey_acronym", "survey_comparability",
    "survey_coverage", "survey_id", "survey_mean_lcu", "survey_mean_ppp",
    "survey_median_lcu", "survey_median_ppp", "survey_time", "survey_year",
    "surveyid_year")
)
}
\arguments{
\item{df}{data.table: Table of results created in \code{fg_pip()}}

\item{cols}{character: Columns with potential duplicate values}
}
\value{
data.table
}
\description{
Remove duplicated rows created during the interpolation process
}
