% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fillin_list}
\alias{fillin_list}
\title{Populate list in parent frame}
\usage{
fillin_list(l, assign = TRUE)
}
\arguments{
\item{l}{list to populate with names objects}

\item{assign}{logical: whether to assign to parent frame}
}
\value{
invisible list \code{l} populated with objects of the same frame
}
\description{
Fill in maned objects of a list with the value of named objects in  the
parent frame in which the list has been created. This objects must have the
same names as the objects of the list
}
\examples{
l <- list(x = NULL,
y = NULL,
z = NULL)

x <-  2
y <-  "f"
z <- TRUE
fillin_list(l)
l
}
