% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lkup.R
\name{filter_lkup}
\alias{filter_lkup}
\title{Filter aggregate distributions when popshare is active}
\usage{
filter_lkup(metadata, popshare)
}
\arguments{
\item{metadata}{data.frame: Output of \code{subset_lkup()}}

\item{popshare}{numeric: popshare value passed to \code{pip()}}
}
\value{
data.frame
}
\description{
The popshare option is not supported for aggregate distributions.
}
\keyword{internal}
