% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_lkup}
\alias{filter_lkup}
\title{Helper to filter metadata
aggregate distribution need to be filtered out when popshare is not null
This is a temporary function until a full fix is implemented, and popshare is
supported for all distributions}
\usage{
filter_lkup(metadata, popshare)
}
\arguments{
\item{metadata}{data.frame: Output of \code{subset_lkup()}}

\item{popshare}{numeric: popshare value passed to \code{pip()}}
}
\value{
data.frame
}
\description{
Helper to filter metadata
aggregate distribution need to be filtered out when popshare is not null
This is a temporary function until a full fix is implemented, and popshare is
supported for all distributions
}
