% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_cumsum.R
\name{format_lfst}
\alias{format_lfst}
\title{format the lfst list to usable data to estimate poverty}
\usage{
format_lfst(lfst, dict)
}
\arguments{
\item{lfst}{List from load_list_refy().}
}
\value{
list with DT and g (GRP object)

List with elements: DT (data.table) and g (GRP object).
}
\description{
format the lfst list to usable data to estimate poverty
}
\keyword{Format}
\keyword{GRP}
\keyword{Takes}
\keyword{a}
\keyword{and}
\keyword{data.table}
\keyword{estimation}
\keyword{for}
\keyword{from}
\keyword{grouped}
\keyword{internal}
\keyword{lfst}
\keyword{list}
\keyword{load_list_refy()}
\keyword{object}
\keyword{operations.}
\keyword{poverty}
\keyword{returns}
\keyword{with}
