% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux_table.R
\name{get_aux_table}
\alias{get_aux_table}
\title{Return specified auxiliary data}
\usage{
get_aux_table(data_dir, data_version, table)
}
\arguments{
\item{data_dir}{character: Data directory}

\item{data_version}{character: Release version of the data in the format yyyymmdd}

\item{table}{character: Name of auxiliary table}
}
\value{
data.frame
}
\description{
Return specified auxiliary data
}
