% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux_table.R
\name{get_aux_table}
\alias{get_aux_table}
\title{Return specified auxiliary data}
\usage{
get_aux_table(data_dir, table)
}
\arguments{
\item{data_dir}{character: Data directory}

\item{table}{character: Name of auxiliary table}
}
\value{

}
\description{
Return specified auxiliary data
}
