% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux_table.R
\name{get_aux_table_ui}
\alias{get_aux_table_ui}
\title{Return specified auxiliary data in wide format
Helper function to the UI}
\usage{
get_aux_table_ui(data_dir, table, exclude = TRUE, lkup)
}
\arguments{
\item{data_dir}{character: Data directory}

\item{table}{character: Name of auxiliary table}

\item{exclude}{logical: whether or not to exclude some countries or regions...}
}
\value{
data.frame
}
\description{
Return specified auxiliary data in wide format
Helper function to the UI
}
\seealso{
Other aux: 
\code{\link{blocked_aux_tables}()},
\code{\link{get_aux_table}()},
\code{\link{reset_blocked_aux_tables}()}
}
\concept{aux}
