% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{get_mean_median}
\alias{get_mean_median}
\title{Merge mean and median from dist_stats into an FGT table}
\usage{
get_mean_median(fgt, lkup, fill_gaps)
}
\arguments{
\item{fgt}{data.table with FGT measures}

\item{lkup}{lkup list}

\item{fill_gaps}{logical: whether lineup-year estimates are being used}
}
\value{
data.table with FGT, mean and median
}
\description{
Early-returns if the lookup is not on the new lineup version.
}
\keyword{internal}
