% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metaregion_table}
\alias{get_metaregion_table}
\title{load metaregion from aux data}
\usage{
get_metaregion_table(data_dir)
}
\arguments{
\item{data_dir}{character: Data directory}
}
\value{
data.table
}
\description{
If there is no data available, return an empty data.frame
}
\keyword{internal}
