% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pg_table}
\alias{get_pg_table}
\title{Load prosperity gap table from aux data}
\usage{
get_pg_table(data_dir, table = c("pg_svy", "pg_lnp"))
}
\arguments{
\item{data_dir}{character: Data directory}

\item{table}{character: Name of auxiliary table}
}
\value{
data.table
}
\description{
If there is no data available, return an empty data.frame
}
\keyword{internal}
