% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pip_version.R
\name{get_pip_version}
\alias{get_pip_version}
\title{Return the versions of the pip packages used for computations}
\usage{
get_pip_version()
}
\value{
list
}
\description{
Return the versions of the pip packages used for computations
}
