% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_spr_table}
\alias{get_spr_table}
\title{load SPR table from aux data}
\usage{
get_spr_table(data_dir, table = c("spr_svy", "spr_lnp"))
}
\arguments{
\item{data_dir}{character: Data directory}

\item{table}{character: Name of auxiliary table}
}
\value{
data.table
}
\description{
If there is no data available, return an empty data.frame
}
\keyword{internal}
