% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_svy_data}
\alias{get_svy_data}
\title{Read survey data}
\usage{
get_svy_data(svy_id, reporting_level, path)
}
\arguments{
\item{svy_id}{character: Survey ID}

\item{reporting_level}{character: geographical reporting level}

\item{path}{character: Path to survey data}
}
\value{
data.frame
}
\description{
Read survey data
}
