% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_aux_long_format_tables}
\alias{get_valid_aux_long_format_tables}
\title{Returns all auxiliary tables that support the long_format=TRUE parameter}
\usage{
get_valid_aux_long_format_tables()
}
\value{
character vector
}
\description{
Returns all auxiliary tables that support the long_format=TRUE parameter
}
