% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_pip.R
\name{load_data_list}
\alias{load_data_list}
\title{load survey year files and store them in a list}
\usage{
load_data_list(metadata)
}
\arguments{
\item{metadata}{data frame from \code{subset_lkup()}}
}
\value{
list with survey years data
}
\description{
load survey year files and store them in a list
}
\keyword{internal}
