% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_func.R
\name{load_inter_cache}
\alias{load_inter_cache}
\title{Load Intermediate cache data}
\usage{
load_inter_cache(lkup = NULL, cache_file_path = NULL, fill_gaps = FALSE)
}
\arguments{
\item{cache_file_path}{file path for cache}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}
}
\value{
data frame
}
\description{
Load Intermediate cache data
}
