% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fgt_new.R
\name{DT_fgt_by_rl}
\alias{DT_fgt_by_rl}
\alias{lt_to_dt}
\alias{map_lt_to_dt}
\alias{map_fgt}
\title{compute FGT using indices by reporting level}
\usage{
DT_fgt_by_rl(x, y, nx, povline)

lt_to_dt(x, y, nx, povline)

map_lt_to_dt(lt, l_rl_rows, povline)

map_fgt(lt, l_rl_rows, povline)
}
\arguments{
\item{x}{data.table from lt list, with welfare and weight vectors}

\item{y}{list of indices for each reporting level}

\item{nx}{name of data table. Usuall country code and year in the form "CCC_YYYY"}

\item{lt}{list of data.tables with welfare and weight data}

\item{l_rl_rows}{list of indeces}
}
\value{
data.table with all measured
}
\description{
This function is intended to be used inside \link{map_fgt}
}
\keyword{internal}
