% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fgt_new.R
\name{map_fgt}
\alias{map_fgt}
\title{map over list of data.tables and indices to compute FGT by reporting_level}
\usage{
map_fgt(lt, l_rl_rows)
}
\arguments{
\item{lt}{list of data.tables with welfare and weight data}

\item{l_rl_rows}{list of indeces}
}
\value{
data.table with all measured
}
\description{
map over list of data.tables and indices to compute FGT by reporting_level
}
\keyword{internal}
