% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{pip}
\alias{pip}
\title{Compute the main PIP poverty and inequality statistics}
\usage{
pip(
  country = "all",
  year = "all",
  povline = 1.9,
  popshare = NULL,
  fill_gaps = FALSE,
  aggregate = FALSE,
  group_by = NULL,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  ppp = NULL,
  lkup
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the poverty line}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate values for missing years}

\item{aggregate}{logical: If set to TRUE, will return aggregate results}

\item{group_by}{character: Will return aggregated values for pre-defined sub-groups}

\item{welfare_type}{character: Welfare type.}

\item{reporting_level}{character: Geographical reporting level}

\item{ppp}{numeric: Custom Purchase Power Parity value}

\item{lkup}{list: A list of lkup tables}
}
\value{
data.frame
}
\description{
Compute the main PIP poverty and inequality statistics
}
