% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{pip}
\alias{pip}
\title{Compute PIP statistics}
\usage{
pip(
  country = "all",
  year = "all",
  povline = 1.9,
  popshare = NULL,
  fill_gaps = FALSE,
  aggregate = FALSE,
  group_by = c("none", "wb"),
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  ppp = NULL,
  lkup,
  debug = FALSE
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{aggregate}{logical: If set to TRUE, will return aggregate results}

\item{group_by}{character: Will return aggregated values for predefined
sub-groups}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{ppp}{numeric: Custom Purchase Power Parity value}

\item{lkup}{list: A list of lkup tables}

\item{debug}{logical: If TRUE poverty calculations from \code{wbpip} will run in
debug mode}
}
\value{
data.table
}
\description{
Compute the main PIP poverty and inequality statistics.
}
\examples{
\dontrun{
# Create lkups
lkups <- create_lkups("<data-folder>")

# A single country and year
pip(country = "AGO",
    year = 2000,
    povline = 1.9,
    lkup = lkups)

# All years for a single country
pip(country = "AGO",
    year = "all",
    povline = 1.9,
    lkup = lkups)

# Fill gaps
pip(country = "AGO",
    year = "all",
    povline = 1.9,
    fill_gaps = TRUE,
    lkup = lkups)

# Group by regions
pip(country = "all",
    year = "all",
    povline = 1.9,
    group_by = "wb",
    lkup = lkups)
}
}
