% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_agg.R, R/pip_grp_logic.R, R/pip_grp_new.R
\name{pip_agg}
\alias{pip_agg}
\alias{pip_grp_logic}
\alias{pip_grp_new}
\title{Logic for computing new aggregate}
\usage{
pip_agg(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  group_by = "wb",
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  lkup,
  censor = FALSE,
  lkup_hash = lkup$cache_data_id$hash_pip_grp,
  additional_ind = FALSE
)

pip_grp_logic(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  group_by = "wb",
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  lkup,
  censor = TRUE,
  lkup_hash = lkup$cache_data_id$hash_pip_grp,
  additional_ind = FALSE
)

pip_grp_new(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  group_by = "wb",
  lkup,
  censor = TRUE,
  additional_ind = FALSE,
  lkup_hash = lkup$cache_data_id$hash_pip_grp
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{group_by}{character: Will return aggregated values for predefined
sub-groups}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{censor}{logical: Triggers censoring of country/year statistics}

\item{lkup_hash}{character: hash of pip}

\item{additional_ind}{logical: If TRUE add new set of indicators. Default if
FALSE}
}
\value{
data.table

data.table
}
\description{
Logic for computing new aggregate

Old way to estimate aggregate data

New way to estimate Aggregate data
}
\examples{
\dontrun{
# Create lkups
}
}
