% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp.R
\name{pip_grp}
\alias{pip_grp}
\title{Compute various aggregations of PIP statistics}
\usage{
pip_grp(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  group_by = c("none", "wb"),
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  lkup,
  censor = TRUE,
  lkup_hash = lkup$cache_data_id$hash_pip_grp
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{group_by}{character: Will return aggregated values for predefined
sub-groups}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{censor}{logical: Triggers censoring of country/year statistics}
}
\value{
data.table
}
\description{
Compute various aggregations of PIP statistics
}
\examples{
\dontrun{
# Create lkups
lkups <- create_lkups("<data-folder>")

# A single country and year
pip_grp(country = "all",
        year = 2000,
        povline = 1.9,
        group_by = "wb",
        lkup = lkups)
}
}
