% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp.R
\name{pip_grp}
\alias{pip_grp}
\title{Compute various aggregations of PIP statistics}
\usage{
pip_grp(
  country = "all",
  year = "all",
  povline = 1.9,
  popshare = NULL,
  group_by = c("none", "wb"),
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  lkup,
  debug = FALSE
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{group_by}{character: Will return aggregated values for predefined
sub-groups}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{debug}{logical: If TRUE poverty calculations from \code{wbpip} will run in
debug mode}
}
\value{
data.table
}
\description{
Compute various aggregations of PIP statistics
}
\examples{
\dontrun{
# Create lkups
lkups <- create_lkups("<data-folder>")

# A single country and year
pip_grp(country = "all",
        year = 2000,
        povline = 1.9,
        group_by = "wb",
        lkup = lkups)
}
}
