% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_grp_logic.R
\name{pip_grp_logic}
\alias{pip_grp_logic}
\title{Logic for computing new aggregate}
\usage{
pip_grp_logic(
  country = "all",
  year = "all",
  povline = 1.9,
  popshare = NULL,
  group_by = c("none", "wb"),
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national"),
  lkup,
  debug = FALSE,
  censor = TRUE
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{group_by}{character: Will return aggregated values for predefined
sub-groups}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{debug}{logical: If TRUE poverty calculations from \code{wbpip} will run in
debug mode}

\item{censor}{logical: Triggers censoring of country/year statistics}
}
\value{
data.table
}
\description{
Logic for computing new aggregate
}
\examples{
\dontrun{
# Create lkups
}
}
