% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_new_lineups.R
\name{pip_new_lineups}
\alias{pip_new_lineups}
\title{Compute PIP statistics}
\usage{
pip_new_lineups(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  popshare = NULL,
  fill_gaps = FALSE,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  ppp = NULL,
  lkup,
  censor = TRUE,
  lkup_hash = lkup$cache_data_id$hash_pip,
  additional_ind = FALSE
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{ppp}{numeric: Custom Purchase Power Parity value}

\item{lkup}{list: A list of lkup tables}

\item{censor}{logical: Triggers censoring of country/year statistics}

\item{lkup_hash}{character: hash of pip}

\item{additional_ind}{logical: If TRUE add new set of indicators. Default if
FALSE}
}
\value{
data.table
}
\description{
Compute the main PIP poverty and inequality statistics.
}
\examples{
\dontrun{
# Create lkups
lkups <- create_lkups("<data-folder>")

# A single country and year
pip_new_lineups(country = "AGO",
    year = 2000,
    povline = 1.9,
    lkup = lkups)

# All years for a single country
pip_new_lineups(country = "AGO",
    year = "all",
    povline = 1.9,
    lkup = lkups)

# Fill gaps
pip_new_lineups(country = "AGO",
    year = "all",
    povline = 1.9,
    fill_gaps = TRUE,
    lkup = lkups)

}
}
