% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipapi-package.R
\docType{package}
\name{pipapi-package}
\alias{pipapi}
\alias{pipapi-package}
\title{pipapi: API for the Poverty and Inequality Platform}
\description{
Provides a high level API to the computations and methods that power the Poverty and Inequality Platform (PIP).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pip-technical-team.github.io/pipapi}
  \item \url{https://github.com/PIP-Technical-Team/pipapi}
  \item Report bugs at \url{https://github.com/PIP-Technical-Team/pipapi/issues}
}

}
\author{
\strong{Maintainer}: R.Andrés Castañeda \email{acastanedaa@worldbank.org}

Authors:
\itemize{
  \item R.Andrés Castañeda \email{acastanedaa@worldbank.org}
  \item Tony Fujs \email{tfujs@worldbank.org}
  \item Aleksander Eilertsen \email{aeilertsen@worldbank.org}
  \item Ronak Shah \email{rshah11@worldbank.org}
}

Other contributors:
\itemize{
  \item World Bank [copyright holder]
}

}
\keyword{internal}
