% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_func.R
\name{return_if_exists}
\alias{return_if_exists}
\title{Return the rows of the table if they exist in master file}
\usage{
return_if_exists(
  lkup,
  povline,
  cache_file_path,
  fill_gaps,
  verbose = getOption("pipapi.verbose")
)
}
\arguments{
\item{lkup}{list: A list of lkup tables}

\item{povline}{numeric: Poverty line}

\item{cache_file_path}{file path for cache}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}
}
\value{
Dataframe
}
\description{
Return the rows of the table if they exist in master file
}
