% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_fun.R
\name{return_if_exists}
\alias{return_if_exists}
\title{Return the rows of the table if they exist in master file}
\usage{
return_if_exists(lkup, is_interpolated, con)
}
\arguments{
\item{lkup}{list: A list of lkup tables}

\item{is_interpolated}{logical : If \code{TRUE}, the call is from \code{fg_pip()}, for \code{FALSE} \code{rg_pip()}}

\item{con}{Connection object}

\item{country_code}{Country Code}
}
\value{
Dataframe
}
\description{
Return the rows of the table if they exist in master file
}
