% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_country}
\alias{select_country}
\title{select_country
Helper function for subset_lkup()}
\usage{
select_country(lkup, keep, country, valid_regions)
}
\arguments{
\item{lkup}{list: A list of lkup tables}

\item{keep}{logical vector}

\item{country}{character: Country ISO 3 codes}

\item{valid_regions}{character: List of valid region codes that can be used
for region selection}
}
\value{
logical vector
}
\description{
select_country
Helper function for subset_lkup()
}
