% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_years}
\alias{select_years}
\title{select_years
Helper function for subset_lkup()}
\usage{
select_years(lkup, keep, year, country)
}
\arguments{
\item{lkup}{list: A list of lkup tables}

\item{keep}{logical vector}

\item{year}{integer: Reporting year}

\item{country}{character: Country ISO 3 codes}
}
\value{
logical vector
}
\description{
select_years
Helper function for subset_lkup()
}
