% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_ctry_years}
\alias{subset_ctry_years}
\title{Subset country-years table
This is a table created at start time to facilitate imputations
It part of the interpolated_list object}
\usage{
subset_ctry_years(country, year, lkup, valid_regions, data_dir)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{lkup}{list: A list of lkup tables}

\item{valid_regions}{character: List of valid region codes that can be used}

\item{data_dir}{character: directory path from lkup$data_root}
}
\value{
data.frame
}
\description{
Subset country-years table
This is a table created at start time to facilitate imputations
It part of the interpolated_list object
}
\keyword{internal}
