% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_ctry_years}
\alias{subset_ctry_years}
\title{Subset country-years table
This is a table created at start time to facilitate imputations
It part of the interpolated_list object}
\usage{
subset_ctry_years(country, year, lkup)
}
\value{
data.frame
}
\description{
Subset country-years table
This is a table created at start time to facilitate imputations
It part of the interpolated_list object
}
\keyword{internal}
