% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_lkup}
\alias{subset_lkup}
\title{Subset look-up data}
\usage{
subset_lkup(
  country,
  year,
  welfare_type,
  reporting_level,
  lkup,
  valid_regions,
  data_dir = NULL,
  povline,
  cache_file_path,
  fill_gaps
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{valid_regions}{character: List of valid region codes that can be used
for region selection}

\item{data_dir}{character: directory path from lkup$data_root}

\item{povline}{numeric: Poverty line}

\item{cache_file_path}{file path for cache}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}
}
\value{
data.frame
}
\description{
Subset look-up data
}
\keyword{internal}
