% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_lkup}
\alias{subset_lkup}
\title{Subset look-up data}
\usage{
subset_lkup(
  country,
  year,
  welfare_type,
  reporting_level,
  lkup,
  valid_regions,
  data_dir = NULL,
  is_interpolated
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{lkup}{list: A list of lkup tables}

\item{valid_regions}{character: List of valid region codes that can be used
for region selection}

\item{data_dir}{character: directory path from lkup$data_root}

\item{is_interpolated}{logical : If \code{TRUE}, the call is from \code{fg_pip()}, for \code{FALSE} \code{rg_pip()}}
}
\value{
data.frame
}
\description{
Subset look-up data
}
\keyword{internal}
