% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipdata.R
\name{transform_input}
\alias{transform_input}
\title{transform input list}
\usage{
transform_input(input_list)
}
\arguments{
\item{input_list}{Character vector of file paths (output from create_full_list).}
}
\value{
formated list
}
\description{
transform input list
}
\keyword{internal}
