% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_cp_charts}
\alias{ui_cp_charts}
\title{Provides numbers that will populate the home page country charts}
\usage{
ui_cp_charts(
  country = "AGO",
  povline = 1.9,
  pop_units = 1e+06,
  year_range = 2016:2019,
  lkup
)
}
\arguments{
\item{country}{character: Country code}

\item{povline}{numeric: Poverty line}

\item{pop_units}{numeric: Units used to express population numbers (default to million)}

\item{year_range}{numeric: Range used to subset countries in Poverty MRV chart}

\item{lkup}{list: A list of lkup tables}
}
\value{
data.frame
}
\description{
Provides numbers that will populate the home page country charts
}
