% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_cp_key_indicators}
\alias{ui_cp_key_indicators}
\title{Country Profiles Key Indicators}
\usage{
ui_cp_key_indicators(country = "AGO", povline = 1.9, lkup)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}
}
\value{
list
}
\description{
Provides numbers that will population for country profiles key indicators.
}
