% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_cp_key_indicators}
\alias{ui_cp_key_indicators}
\title{Provides numbers that will population key indicators for Country Profiles}
\usage{
ui_cp_key_indicators(country = "AGO", povline = NULL, lkup)
}
\arguments{
\item{country}{character: Country code}

\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}
}
\value{
data.frame
}
\description{
Provides numbers that will population key indicators for Country Profiles
}
