% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_cp_poverty_charts}
\alias{ui_cp_poverty_charts}
\title{Provides numbers that will populate the country profiles poverty charts}
\usage{
ui_cp_poverty_charts(country, povline, pop_units, year_range, lkup)
}
\arguments{
\item{country}{character: Country code}

\item{povline}{numeric: Poverty line}

\item{pop_units}{numeric: Units used to express population numbers (default to million)}

\item{year_range}{numeric: Range used to subset countries in Poverty MRV chart}

\item{lkup}{list: A list of lkup tables}
}
\value{
list
}
\description{
Provides numbers that will populate the country profiles poverty charts
}
\keyword{internal}
