% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{ui_hp_stacked}
\alias{ui_hp_stacked}
\title{Provides numbers that will populate the home page main chart}
\usage{
ui_hp_stacked(povline = 1.9, lkup, paths)
}
\arguments{
\item{povline}{numeric: Poverty line}

\item{lkup}{list: A list of lkup tables}

\item{paths}{character: Path to input data}
}
\value{
data.frame
}
\description{
Provides numbers that will populate the home page main chart
}
