% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_poverty_indicators.R
\name{ui_pc_charts}
\alias{ui_pc_charts}
\title{Poverty Calculator Main chart}
\usage{
ui_pc_charts(
  country = c("AGO"),
  year = "all",
  povline = 1.9,
  fill_gaps = FALSE,
  group_by = "none",
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  pop_units = 1e+06,
  censor = TRUE,
  lkup,
  lkup_hash = lkup$cache_data_id$hash_pip_grp
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}

\item{welfare_type}{character: Welfare type}

\item{reporting_level}{character: Geographical reporting level}

\item{pop_units}{numeric: Units used to express population numbers (default
to million)}

\item{censor}{logical: Triggers censoring of country/year statistics}

\item{lkup}{list: A list of lkup tables}

\item{lkup_hash}{character: hash of pip}
}
\value{
data.table
}
\description{
Provides numbers that will populate the poverty calculator main chart.
}
