% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_poverty_indicators.R
\name{ui_pc_regional}
\alias{ui_pc_regional}
\title{Poverty Calculator regional aggregates}
\usage{
ui_pc_regional(
  country = "ALL",
  year = "ALL",
  povline = 1.9,
  pop_units = 1000000,
  lkup
)
}
\arguments{
\item{country}{character: Country ISO 3 codes}

\item{year}{integer: Reporting year}

\item{povline}{numeric: Poverty line}

\item{pop_units}{numeric: Units used to express population numbers (default
to million)}

\item{lkup}{list: A list of lkup tables}
}
\value{
data.table
}
\description{
Provides numbers that will populate poverty calculator regional aggregates
for all years.
}
