% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_func.R
\name{update_master_file}
\alias{update_master_file}
\title{Update master file with the contents of the dataframe}
\usage{
update_master_file(
  dat,
  cache_file_path,
  fill_gaps,
  verbose = getOption("pipapi.verbose"),
  decimal = 2
)
}
\arguments{
\item{dat}{Dataframe to be appended}

\item{cache_file_path}{path where cache file is saved}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}
}
\value{
number of rows updated
}
\description{
Update master file with the contents of the dataframe
}
