% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_fun.R
\name{update_master_file}
\alias{update_master_file}
\title{Update master file with the contents of the dataframe}
\usage{
update_master_file(dat, con, fill_gaps)
}
\arguments{
\item{dat}{Dataframe to be appended}

\item{con}{DuckDB connection object}

\item{fill_gaps}{logical: If set to TRUE, will interpolate / extrapolate
values for missing years}
}
\value{
number of rows updated
}
\description{
Update master file with the contents of the dataframe
}
